refactor(interface): modularize components >300 lines and audit exhaustive-deps
- Refactored 12 monolithic UI components in apps/interface/src/components down to <300 lines by extracting custom hooks and modular presentation subcomponents.
- Extracted business logic into custom React hooks (useTeamManagement.ts, useEmbedCodeGenerator.ts) to separate data management from UI layout.
- Modularized component rendering into feature-based subcomponents (team, search, referral, embed, leaderboard, achievement, comment, transaction, wizard, export).
- Audited all react-hooks/exhaustive-deps suppressions across TeamManagement.tsx, useAsync.ts, useCampaignFilters.ts, useAdvancedSearch.ts, and CampaignsSection.tsx, resolving dependency arrays with useCallback or documenting explicit safety rationale.
- Explicitly documented TeamManagement.tsx mock-data status and backend API requirements in docs/TEAM_MANAGEMENT_BACKEND_PREREQUISITE.md.
