+++ apps/interface/src/components/gamification/Leaderboard.tsx
@@ -1,22 +1,12 @@
 "use client";
 
 import React, { useState } from "react";
-import { motion, AnimatePresence } from "framer-motion";
-import {
-  Trophy,
-  ChevronLeft,
-  ChevronRight,
-  Medal,
-  TrendingUp,
-  Eye,
-  EyeOff,
-} from "lucide-react";
+import { AnimatePresence } from "framer-motion";
+import { Trophy, ChevronLeft, ChevronRight } from "lucide-react";
 import { cn } from "@/lib/utils";
-import type {
-  LeaderboardEntry,
-  LeaderboardFilter,
-  ChallengeLeaderboardEntry,
-} from "@/types/gamification";
+import type { LeaderboardEntry } from "@/types/gamification";
+import { LeaderboardRow } from "./leaderboard/LeaderboardRow";
+import { LeaderboardControls } from "./leaderboard/LeaderboardControls";
 
 interface LeaderboardProps {
   entries: LeaderboardEntry[];
@@ -32,139 +22,6 @@ interface LeaderboardProps {
   pageSize?: number;
 }
 
-interface LeaderboardRowProps {
-  entry: LeaderboardEntry;
-  isCurrentUser: boolean;
-  showAddresses: boolean;
-}
-
-// Medal icons for top 3
-const medalIcons = {
-  1: <Medal className="w-5 h-5 text-yellow-500" />,
-  2: <Medal className="w-5 h-5 text-gray-400" />,
-  3: <Medal className="w-5 h-5 text-orange-400" />,
-};
-
-/**
- * Format address for display (shortened)
- */
-function formatAddress(address: string, show: boolean = true): string {
-  if (!show) return "••••••••";
-  return `${address.slice(0, 6)}...${address.slice(-4)}`;
-}
-
-/**
- * Leaderboard row component
- */
-function LeaderboardRow({
-  entry,
-  isCurrentUser,
-  showAddresses,
-}: LeaderboardRowProps) {
-  const isTopThree = entry.rank <= 3;
-  const medal = medalIcons[entry.rank as 1 | 2 | 3];
-
-  return (
-    <motion.tr
-      initial={{ opacity: 0, x: -20 }}
-      animate={{ opacity: 1, x: 0 }}
-      exit={{ opacity: 0, x: -20 }}
-      transition={{ duration: 0.2 }}
-      className={cn(
-        "border-b border-gray-200 dark:border-gray-700 last:border-0 transition-colors",
-        "hover:bg-gray-50 dark:hover:bg-gray-800/50",
-        isCurrentUser && "bg-blue-50 dark:bg-blue-900/20",
-      )}
-    >
-      {/* Rank */}
-      <td className="px-4 py-3">
-        <div className="flex items-center gap-2">
-          {isTopThree ? (
-            medal
-          ) : (
-            <span className="text-sm font-semibold text-gray-600 dark:text-gray-400 w-5">
-              {entry.rank}
-            </span>
-          )}
-        </div>
-      </td>
-
-      {/* Display name / Address */}
-      <td className="px-4 py-3">
-        <div className="flex items-center gap-2">
-          {entry.displayName ? (
-            <>
-              <span className="font-medium text-gray-900 dark:text-white">
-                {entry.displayName}
-              </span>
-              <span className="text-xs text-gray-500 dark:text-gray-400">
-                {formatAddress(entry.address, showAddresses)}
-              </span>
-            </>
-          ) : (
-            <span className="font-medium text-gray-900 dark:text-white">
-              {formatAddress(entry.address, showAddresses)}
-            </span>
-          )}
-          {isCurrentUser && (
-            <span className="text-xs bg-blue-100 dark:bg-blue-900 text-blue-700 dark:text-blue-300 px-2 py-0.5 rounded-full">
-              You
-            </span>
-          )}
-        </div>
-      </td>
-
-      {/* Level badge */}
-      <td className="px-4 py-3">
-        <div className="flex items-center gap-2">
-          <div className="w-8 h-8 rounded-full bg-gradient-to-br from-purple-400 to-purple-600 flex items-center justify-center text-white text-xs font-bold">
-            {entry.level}
-          </div>
-          <span className="text-sm text-gray-600 dark:text-gray-400">Lvl</span>
-        </div>
-      </td>
-
-      {/* Points / Score */}
-      <td className="px-4 py-3 text-right">
-        <div className="font-semibold text-gray-900 dark:text-white">
-          {entry.totalPoints.toLocaleString()}
-        </div>
-        <div className="text-xs text-gray-500 dark:text-gray-400">points</div>
-      </td>
-
-      {/* Contributions count */}
-      <td className="px-4 py-3 text-right">
-        <div className="font-semibold text-gray-900 dark:text-white">
-          {entry.contributionCount}
-        </div>
-        <div className="text-xs text-gray-500 dark:text-gray-400">
-          contribution{entry.contributionCount !== 1 ? "s" : ""}
-        </div>
-      </td>
-
-      {/* Achievements */}
-      <td className="px-4 py-3 text-center">
-        <div className="inline-flex items-center gap-2 bg-yellow-100 dark:bg-yellow-900/30 text-yellow-700 dark:text-yellow-300 px-3 py-1 rounded-full text-sm font-medium">
-          <Trophy size={14} />
-          {entry.achievements}
-        </div>
-      </td>
-
-      {/* Badge */}
-      {entry.badge && (
-        <td className="px-4 py-3">
-          <span className="text-sm px-3 py-1 bg-gradient-to-r from-blue-500 to-purple-600 text-white rounded-full">
-            ⭐ {entry.badge}
-          </span>
-        </td>
-      )}
-    </motion.tr>
-  );
-}
-
-/**
- * Main Leaderboard Component
- */
 export function Leaderboard({
   entries,
   userAddress,
@@ -179,13 +36,6 @@ export function Leaderboard({
   pageSize = 10,
 }: LeaderboardProps) {
   const [showAddresses, setShowAddresses] = useState(false);
-
-  const timeframeOptions: Array<
-    "all-time" | "this-month" | "this-week"
-  > = ["all-time", "this-month", "this-week"];
-  const typeOptions: Array<
-    "points" | "contributions" | "achievements" | "referrals"
-  > = ["points", "contributions", "achievements", "referrals"];
 
   const typeLabels = {
     points: "Top Points",
@@ -227,65 +77,14 @@ export function Leaderboard({
         </div>
       </div>
 
-      {/* Controls */}
-      <div className="flex flex-col sm:flex-row gap-4 items-start sm:items-center justify-between">
-        {/* Timeframe tabs */}
-        <div className="flex gap-2 border-b border-gray-200 dark:border-gray-700">
-          {timeframeOptions.map((tf) => (
-            <button
-              key={tf}
-              onClick={() => onTimeframeChange?.(tf)}
-              className={cn(
-                "px-4 py-2 font-medium text-sm transition-colors capitalize",
-                timeframe === tf
-                  ? "text-yellow-600 dark:text-yellow-400 border-b-2 border-yellow-600 dark:border-yellow-400"
-                  : "text-gray-600 dark:text-gray-400 hover:text-gray-900 dark:hover:text-gray-200",
-              )}
-            >
-              {tf.replace(/-/g, " ")}
-            </button>
-          ))}
-        </div>
-
-        {/* Type buttons */}
-        <div className="flex gap-2 flex-wrap">
-          {typeOptions.map((t) => (
-            <button
-              key={t}
-              onClick={() => onTypeChange?.(t)}
-              className={cn(
-                "px-3 py-1 rounded-full text-sm font-medium transition-colors capitalize",
-                type === t
-                  ? "bg-yellow-500 text-white"
-                  : "bg-gray-100 dark:bg-gray-800 text-gray-700 dark:text-gray-300 hover:bg-gray-200 dark:hover:bg-gray-700",
-              )}
-            >
-              {t}
-            </button>
-          ))}
-        </div>
-
-        {/* Toggle address visibility */}
-        <button
-          onClick={() => setShowAddresses(!showAddresses)}
-          className="flex items-center gap-2 text-sm text-gray-500 dark:text-gray-400 hover:text-gray-900 dark:hover:text-white transition"
-          aria-label={
-            showAddresses ? "Hide addresses" : "Show addresses"
-          }
-        >
-          {showAddresses ? (
-            <>
-              <Eye size={16} />
-              <span>Hide</span>
-            </>
-          ) : (
-            <>
-              <EyeOff size={16} />
-              <span>Show</span>
-            </>
-          )}
-        </button>
-      </div>
+      <LeaderboardControls
+        timeframe={timeframe}
+        type={type}
+        showAddresses={showAddresses}
+        onTimeframeChange={onTimeframeChange}
+        onTypeChange={onTypeChange}
+        onToggleAddresses={() => setShowAddresses(!showAddresses)}
+      />
 
       {/* Leaderboard table */}
       <div className="overflow-x-auto rounded-lg border border-gray-200 dark:border-gray-700 bg-white dark:bg-gray-900">
